requirements.txt
langchain>=0.1.0
langgraph>=0.0.40
langchain-openai>=0.0.5
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pydantic>=2.5.0
tavily-python>=0.3.0
supabase>=2.0.0
matplotlib>=3.7.0
pandas>=2.0.0
streamlit>=1.29.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

